+++ frontend/pages/index.module.scss
@@ -1,4 +1,4 @@
-@use '../../styles/customVariables' as c-vars;
+@use '../styles/customVariables' as c-vars;
 
 .main {
   min-height: 100vh;
